+++ analyzers/__init__.py
@@ -5,9 +5,9 @@ from .example.example_report import example_report
 from .example.example_web import example_web
 from .hashtags import hashtags
 from .hashtags_web import hashtags_web
-from .ngram_stats import ngram_stats
-from .ngram_web import ngrams_web
-from .ngrams import ngrams
+from .ngrams.ngram_stats import ngram_stats
+from .ngrams.ngram_web import ngrams_web
+from .ngrams.ngrams_base import ngrams
 from .temporal import temporal
 from .temporal_barplot import temporal_barplot
 from .time_coordination import time_coordination
