+++ analyzers/ngrams/ngram_stats/interface.py
@@ -1,7 +1,7 @@
 from analyzer_interface import AnalyzerOutput, OutputColumn, SecondaryAnalyzerInterface
 
-from ..ngrams import interface as ngrams_interface
-from ..ngrams.interface import (
+from ..ngrams_base import interface as ngrams_interface
+from ..ngrams_base.interface import (
     COL_AUTHOR_ID,
     COL_MESSAGE_ID,
     COL_MESSAGE_NGRAM_COUNT,
